requirements.txt
pytest-playwright==0.7.2
playwright==1.58.0
python-dotenv==1.0.1
pytest==9.0.2
